Moved a function into `AppStorage`.
